+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+## 1.0.3
+* Added `Level` to control logging output.
+
 ## 1.0.2
 * Added `PrettyOutputType` to allow for custom output types.
-*
+
 ## 1.0.1
-* fix publish warnings
+* fix publish warnings.
 
 ## 1.0.0
 * Initial version.